offers/requirements.txt
aiofiles==0.6.0
aiopg==1.0.0
asyncpg==0.21.0
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.3
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
python-dotenv==0.15.0
rfc3986==1.4.0
sanic==20.9.1
sanic-openapi==0.6.2
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.20
ujson==4.0.1
uvloop==0.14.0
websockets==8.1
